// DXC_Frontend/src/api/models/insertQuestionCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface InsertQuestionCommand {
  surveyId?: number
  /** @nullable */
  noiDung: string | null
  /** @nullable */
  cauHoiTuLuan?: string | null
  /** @nullable */
  stt?: number | null
}