// DXC_Frontend/src/api/models/pagedResultOfEssayQuestionDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { EssayQuestionDto } from './essayQuestionDto'

/**
 * Represents a paginated API response.
 */
export interface PagedResultOfEssayQuestionDto {
  /** Indicates whether the request was successful. */
  success?: boolean
  /**
   * The list of data items for the current page.
   * @nullable
   */
  data?: EssayQuestionDto[] | null
  /** The total number of records. */
  total?: number
  /** The current page number. */
  current?: number
  /** The number of items per page. */
  pageSize?: number
  /**
   * An optional message, typically used for errors.
   * @nullable
   */
  message?: string | null
}